+++ test/ThreadedTest.cpp
@@ -64,12 +64,12 @@ int main() {
 			vec->push(SObject{100, "Thread One"});
 			vec->top()->print();
 			//vec->doFor(0, [](const TUnique<SObject>& obj) { obj->print(); });
-			vec->pop(0);
+			vec->pop(static_cast<size_t>(0));
 		});
 		vec->push(SObject{101, "Thread Two"});
 		vec->top()->print();
 		//vec->doFor(0, [](TUnique<SObject>& obj) { obj->print(); });
-		vec->pop(0);
+		vec->pop(static_cast<size_t>(0));
 
 		std::cout << "Continue?" << std::endl;
 		std::cin >> c;
